DEBUG: added temporary message

+++ host/examples/rx_stream_init.cpp
@@ -91,7 +91,10 @@ template<typename samp_type> void recv_to_file(
         std::this_thread::sleep_for(std::chrono::microseconds((int)(time_requested*1e6)));
     }
 
-    kill(pre_pid, SIGTERM);
+    if(pre_pid != 0) {
+        std::cout << "Stopping pre exec" <<std::endl;
+        kill(pre_pid, SIGTERM);
+    }
 
     stream_cmd.stream_mode = uhd::stream_cmd_t::STREAM_MODE_STOP_CONTINUOUS;
     rx_stream->issue_stream_cmd(stream_cmd);
